tests/fastapi-framework/fastapi-framework/requirements.txt
# Core
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
uvloop>=0.21.0
httptools>=0.6.0
orjson>=3.10.12

# Database
sqlmodel>=0.0.21
asyncpg>=0.30.0
psycopg2-binary>=2.9.10
alembic>=1.14.0

# Redis
redis>=5.2.0
hiredis>=3.1.0

# Auth
python-jose[cryptography]>=3.3.0
bcrypt>=4.2.1

# Utils
python-multipart>=0.0.20
pillow>=11.1.0
pydantic-settings>=2.7.0
httpx>=0.28.1

# Cache
cachetools>=5.5.0

# Task queue
arq>=0.24.1

# S3 (optional — comment out if not using object storage)
aiobotocore>=2.13.0
boto3>=1.34.0

# Monitoring
prometheus-client>=0.19.0
